+++ src/Akeneo/Pim/Enrichment/AssetManager/Component/Updater/Copier/AssetCollectionAttributeCopier.php
@@ -1,0 +1,101 @@
+<?php
+declare(strict_types=1);
+
+/*
+ * This file is part of the Akeneo PIM Enterprise Edition.
+ *
+ * (c) 2018 Akeneo SAS (http://www.akeneo.com)
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Akeneo\Pim\Enrichment\AssetManager\Component\Updater\Copier;
+
+use Akeneo\Pim\Enrichment\Component\Product\Builder\EntityWithValuesBuilderInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Model\EntityWithValuesInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Updater\Copier\AbstractAttributeCopier;
+use Akeneo\Pim\Enrichment\Component\Product\Validator\AttributeValidatorHelper;
+use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
+
+/**
+ * Value copier for Asset Collection attributes
+ *
+ * @author    Nicolas Maligne <nicolas.maligne@akeneo.com>
+ * @copyright 2018 Akeneo SAS (https://www.akeneo.com)
+ */
+class AssetCollectionAttributeCopier extends AbstractAttributeCopier
+{
+    public function __construct(
+        EntityWithValuesBuilderInterface $entityWithValuesBuilder,
+        AttributeValidatorHelper $attrValidatorHelper,
+        array $supportedFromTypes,
+        array $supportedToTypes
+    ) {
+        parent::__construct($entityWithValuesBuilder, $attrValidatorHelper);
+
+        $this->supportedFromTypes = $supportedFromTypes;
+        $this->supportedToTypes = $supportedToTypes;
+    }
+
+    public function supportsAttributes(AttributeInterface $fromAttribute, AttributeInterface $toAttribute)
+    {
+        $supportsFrom = in_array($fromAttribute->getType(), $this->supportedFromTypes);
+        $supportsTo = in_array($toAttribute->getType(), $this->supportedToTypes);
+        $sameAssetFamily = ($fromAttribute->getReferenceDataName() === $toAttribute->getReferenceDataName());
+
+        return $supportsFrom && $supportsTo &&  $sameAssetFamily;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function copyAttributeData(
+        EntityWithValuesInterface $fromEntityWithValues,
+        EntityWithValuesInterface $toEntityWithValues,
+        AttributeInterface $fromAttribute,
+        AttributeInterface $toAttribute,
+        array $options = []
+    ) {
+        $options = $this->resolver->resolve($options);
+        $fromLocale = $options['from_locale'];
+        $toLocale = $options['to_locale'];
+        $fromScope = $options['from_scope'];
+        $toScope = $options['to_scope'];
+
+        $this->checkLocaleAndScope($fromAttribute, $fromLocale, $fromScope);
+        $this->checkLocaleAndScope($toAttribute, $toLocale, $toScope);
+
+        $this->copySingleValue(
+            $fromEntityWithValues,
+            $toEntityWithValues,
+            $fromAttribute,
+            $toAttribute,
+            $fromLocale,
+            $toLocale,
+            $fromScope,
+            $toScope
+        );
+    }
+
+    protected function copySingleValue(
+        EntityWithValuesInterface $fromEntityWithValues,
+        EntityWithValuesInterface $toEntityWithValues,
+        AttributeInterface $fromAttribute,
+        AttributeInterface $toAttribute,
+        ?string $fromLocale,
+        ?string $toLocale,
+        ?string $fromScope,
+        ?string $toScope
+    ): void {
+        $fromValue = $fromEntityWithValues->getValue($fromAttribute->getCode(), $fromLocale, $fromScope);
+        $value = (null !== $fromValue && null !== $fromValue->getData()) ? $fromValue->getData() : null;
+        $this->entityWithValuesBuilder->addOrReplaceValue(
+            $toEntityWithValues,
+            $toAttribute,
+            $toLocale,
+            $toScope,
+            $value
+        );
+    }
+}